Add the global output
The new global output is status.36
The detailed output data is written in comments in global_params_gether.f and global_output.F
The nb6.py and nb6_status.ipynb are used to read the data.
nb6.py include a function to read the data and nb6_status.ipynb is the Jupyter notebook to show how to use this function and how to draw the data.

The current version is for testing. Many output data need to be check.